new branch / modify project for host

+++ php/del_image.php
@@ -1,15 +1,23 @@
 <?php 
 	session_start();
 	if(empty($_SESSION['auth'])){
-		header("Location:" . $_SERVER['REQUEST_SCHEME'] . '://' . $_SERVER['HTTP_HOST'] . '/AurelBs/index.php');
+		header("Location:" . $_SERVER['HTTP_X_FORWARDED_PROTO'] . '://' . $_SERVER['HTTP_HOST'] . '/index.php');
 	}
 
-	if (!empty($_GET['idi'])) {
+	if (!empty($_GET['idi']) && !empty($_GET['idg'])) {
 		require_once("connect.php");
 		$idi = $_GET['idi'];
+		$idg = $_GET['idg'];
 		$sql = "DELETE FROM `images` WHERE `images`.`id` = $idi";
 		$con->query($sql);
 
-		header("Location:" . $_SERVER['REQUEST_SCHEME'] . '://' . $_SERVER['HTTP_HOST'] . '/AurelBs/index.php');
+		if (!empty($_GET['idc'])) {
+			$idc = $_GET['idc'];
+			header("Location:" . $_SERVER['HTTP_X_FORWARDED_PROTO'] . '://' . $_SERVER['HTTP_HOST'] . '/single.php?idg=' . $idg . '&idc=' . $idc);
+		}else {
+			$idc = "";
+		}
+
+		header("Location:" . $_SERVER['HTTP_X_FORWARDED_PROTO'] . '://' . $_SERVER['HTTP_HOST'] . '/single.php?idg=' . $idg . '&idc=' . $idc);
 	}
  ?>
